+++ Changelog.md
@@ -6,11 +6,13 @@
  - added new `Parameters.null_string`
  - added new unsigned `Parameters.ubigid` and `Parameters.uid`
  - WARNING: no longer accepts `nil` for all parameters, use ` | Parameters.nil`, for transition period you can use this to log all calls that would result in `StrongerParameters::InvalidParameter`
+ - `action_on_invalid_parameters` will now raise by default
+ - `action_on_invalid_parameters` is now set on `ActionController::Parameters`
 
 ```Ruby
 # tested in test/unit/zendesk/extensions/strong_parameters_test.rb
 # remove once we no longer see log messages in production
-config.action_controller.action_on_invalid_parameters = lambda do |result, key|
+ActionController::Parameters.action_on_invalid_parameters = lambda do |result, key|
   if result.value.nil? && Rails.env.production?
     Rails.logger.error("Previously allowed nil value passed #{key} -- #{result.message}")
   else